requirements.txt
Django
djangorestframework
celery
Redis
psycopg2-binary